- refactoring of hood light and fan code
