+++ custom_components/ge_home/entities/common/__init__.py
@@ -7,6 +7,7 @@ from .ge_cooktop_status_binary_sensor import GeCooktopStatusBinarySensor
 from .ge_erd_property_binary_sensor import GeErdPropertyBinarySensor
 from .ge_erd_sensor import GeErdSensor
 from .ge_erd_light import GeErdLight
+from .ge_erd_fan import GeErdFan
 from .ge_erd_timer_sensor import GeErdTimerSensor
 from .ge_erd_property_sensor import GeErdPropertySensor
 from .ge_erd_switch import GeErdSwitch
