+++ custom_components/ge_home/entities/hood/__init__.py
@@ -1,2 +1,4 @@
-from .ge_hood_fan_speed import GeHoodFan, GeHoodFanSpeedSelect
-from .ge_hood_light_level import GeHoodLight, GeHoodLightLevelSelect
+from .ge_hood_fan import GeHoodFan
+from .ge_hood_fan_speed import GeHoodFanSpeedSelect
+from .ge_hood_light import GeHoodLight
+from .ge_hood_light_level import GeHoodLightLevelSelect
